in case of ERROR_NETNAME_DELETED , ignore and retry to accept a new connection
as i have found in cpython issues and other similar docs:

Winsock special cases several status codes returned by the AFD device that implements socket files. For example:

| NTSTATUS                 | Winsock Error   | Generic Windows Error |
|--------------------------|-----------------|-----------------------|
| STATUS_LOCAL_DISCONNECT  | WSAECONNABORTED | ERROR_NETNAME_DELETED |
| STATUS_CONNECTION_RESET  | WSAECONNRESET   | ERROR_NETNAME_DELETED |
| STATUS_REMOTE_DISCONNECT | WSAECONNRESET   | ERROR_NETNAME_DELETED |

Either way it means the connection was reset or disconnected, so it shuold ignore this failure and retry for accepting new connections.