avoid double `GC_unref` when final `boundstream.transfer` is cancelled
When a `BoundedStreamReader` finishes, it performs one final `transfer`
to signal EOF. If the other end subsequently closes the stream, state
changes from `Finished` to `Closing` and enqueues a callback that will
finish the close operation. If the `transfer` is then cancelled, the
`state` changes from `Closing` to `Error`. The stream will eventually
be closed, so changes again from `Error` to `Closing`, and the cleanup
callback is scheduled a second time. Each of the cleanup callbacks then
proceeds to `GC_unref` to `udata` before changing the state to `Closed`.
The double `GC_unref` only happens if `udata` is used.

In practice, the buggy flow is triggered when `nimbus-eth2` interacts
with Geth, and Geth provides a response with `Content-Length`. This is
the case for example when a block without transactions is fetched;
otherwise, it uses `Transfer-Encoding: chunked`. Luckily, that setup
does not configure `udata`, so there is no double free in practice.
But, other applications that use `udata` and use the cancellation flow
that triggers the bug may be affected.